recipe/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bootstrap4==0.1.0
boto3==1.28.5
botocore==1.31.5
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
crispy-bootstrap4==2022.1
distlib==0.3.7
Django==4.2.3
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-storages==1.13.2
file-validator==1.0.9
filelock==3.12.2
filetype==1.1.0
frozenlist==1.4.0
humanize==4.4.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
openai==0.28.0
Pillow==10.0.0
platformdirs==3.10.0
psycopg2==2.9.7
puremagic==1.14
python-dateutil==2.8.2
python-dotenv==0.21.1
python-magic==0.4.27
python-magic-bin==0.4.14
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
termcolor==1.1.0
tqdm==4.66.1
tzdata==2023.3
urllib3==1.26.16
whitenoise==6.5.0
yarl==1.9.2
